Switches amd64 runner with a self-hosted one
There are some issues when trying to use the github runners for
building the envoy rock, specifically when it comes to building
and linking the envoy binary. The Runner can lose communication
with the server due to CPU / Memory starvation, causing the runner
to be terminated.

The self-hosted runners are bigger, so the compilation time should
be significantly smaller, and it shouldn't end up in resource starvation.